+++ TombLauncher/Services/GameWithStatsService.cs
@@ -250,4 +250,34 @@ public class GameWithStatsService : IViewService
 
         process.Start();
     }
+
+    public async Task ToggleFavourite(GameWithStatsViewModel gameWithStatsViewModel)
+    {
+        var metadata = _mapper.Map<GameMetadataDto>(gameWithStatsViewModel.GameMetadata);
+        metadata.IsFavourite = !metadata.IsFavourite;
+        await _gamesUnitOfWork.UpsertGame(metadata);
+        gameWithStatsViewModel.GameMetadata.IsFavourite = metadata.IsFavourite;
+    }
+
+    public async Task ToggleCompleted(GameWithStatsViewModel gameWithStatsViewModel)
+    {
+        var metadata = _mapper.Map<GameMetadataDto>(gameWithStatsViewModel.GameMetadata);
+        metadata.IsCompleted = !metadata.IsCompleted;
+        await _gamesUnitOfWork.UpsertGame(metadata);
+        gameWithStatsViewModel.GameMetadata.IsCompleted = metadata.IsCompleted;
+    }
+    
+    public bool CanUninstall(GameMetadataViewModel metadataViewModel)
+    {
+        return metadataViewModel.IsInstalled;
+    }
+
+    public async Task Uninstall(string installDir, int gameId)
+    {
+        NavigationManager.GetCurrentPage().SetBusy("Uninstalling...");
+        Directory.Delete(installDir, true);
+        _gamesUnitOfWork.MarkGameAsUninstalled(gameId);
+        await _gamesUnitOfWork.Save();
+        NavigationManager.GoBack();
+    }
 }